Add station status feature and integrate with UI
Introduces StationStatusDto, IStationStatusService, and StationStatusService for retrieving and displaying station status information. Updates dependency injection, PLC client, and snapshot reader to support new status fields. Refactors WinForms HomePage and status bar controls to render station status, and updates configuration and entity types for compatibility.

+++ ISKI.IBKS.Domain/Entities/StationSnapshot.cs
@@ -9,11 +9,11 @@ namespace ISKI.IBKS.Domain.Entities;
 public class StationSnapshot
 {
     public DateTime? SystemTime { get; set; }
-    public short? HaftalikYikamaGunu { get; set; }
-    public short? HaftalikYikamaSaati { get; set; }
-    public short? SaatlikYikamaSaati { get; set; }
-    public short? YikamaDakikasi { get; set; }
-    public short? YikamaSaniyesi { get; set; }
+    public int? HaftalikYikamaGunu { get; set; }
+    public int? HaftalikYikamaSaati { get; set; }
+    public int? SaatlikYikamaSaati { get; set; }
+    public int? YikamaDakikasi { get; set; }
+    public int? YikamaSaniyesi { get; set; }
     public double? TesisDebi { get; set; }
     public double? TesisGunlukDebi { get; set; }
     public double? DesarjDebi { get; set; }
